requirements.txt
pytube
moviepy
openai
requests
python-dotenv
flask-cors